requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
streamlit
pandas
openai
python-telegram-bot